move sidekiq integration to applicaster/logger/sidekiq
