+++ lib/applicaster-logger.rb
@@ -1,5 +1,3 @@
 require 'applicaster/logger'
 require "applicaster/rack/request_data"
-require "applicaster/sidekiq/middleware"
-require "applicaster/sidekiq/job_logger"
-require "applicaster/sidekiq/exception_logger"
+require "applicaster/logger/sidekiq"
