+++ lib/applicaster/logger/sidekiq/job_logger.rb
